macOS/iOS: bugfix for devices exposing secondary services with no corresponding primary service
A device can expose services that are marked as secondary, but not exposing primary services including them. This causes the service for a characteristic to become null and triggers exceptions on the dart side.

This fix adds a fallback behavior that uses the secondary service as the primary if another primary is not found